build: Disable noisy unused resources check
So far lint lacks the ability to detect resources that are
only being used via databinding and thus do not have a direct
R.java reference being used in code, making for a lot of false
positives in our databinding driven codebase. Disable the check
until this changes, if ever.

Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>