Ensure we don't clear existing dependencies
When we added the spotlessCheck dependency to assemble tasks,
we ignored the existing dependencies of the task and overwrote
them, which caused assemble tasks to simply run spotless and exit
since it didn't know what else to do. Add all existing tasks
to our dependencies list before calling the setter to ensure
this doesn't happen.

Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>